requirements.txt
openai
streamlit
langchain
chromadb
tqdm
python-dotenv
